requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
diskcache==5.6.3
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
jinja2==3.1.3
llama-cpp-python==0.2.47
markupsafe==2.1.5
numpy==1.26.4
packaging==23.2
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.2
typing-extensions==4.9.0
urllib3==2.2.1
